requirements.txt
Flask
requests
openai
langchain
faiss-cpu
tiktoken
gunicorn
python-dotenv